python/pyproject.toml
[project]
name = "redbull-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "psycopg2>=2.9.11",
    "textual-dev>=1.8.0",
]
